Move GLTF files into own project #2
