+++ indra/llgltf/CMakeLists.txt
@@ -5,7 +5,7 @@ project(llgltf)
 include(00-Common)
 include(LLCommon)
 include(LLMath)
-include(LLPrimitive)
+include(LLPhysicsExtensions)
 include(GLM)
 
 set(llgltf_SOURCE_FILES
@@ -14,6 +14,9 @@ set(llgltf_SOURCE_FILES
     gltfprimitive/asset.cpp
     gltfprimitive/primitive.cpp
     gltfprimitive/llgltfloader.cpp
+    llgltfmaterialcore.cpp
+    llmodel.cpp
+    llmodelloader.cpp
     )
 
 set(llgltf_HEADER_FILES
@@ -26,6 +29,9 @@ set(llgltf_HEADER_FILES
     gltfprimitive/primitive.h
     gltfprimitive/llgltfloader.h
     gltfprimitive/buffer_util.h
+    llgltfmaterialcore.h
+    llmodel.h
+    llmodelloader.h
     )
 
 list(APPEND llgltf_SOURCE_FILES ${llgltf_HEADER_FILES})
@@ -41,9 +47,30 @@ target_include_directories(llgltf PUBLIC
 target_link_libraries(llgltf
     llcommon
     llmath
-    llprimitive
+    llappearance
+    llrender
     )
+
+if (HAVOK OR HAVOK_TPV)
+    target_link_libraries(llgltf
+        llphysicsextensions_impl
+        )
+else()
+    target_link_libraries(llgltf
+        llphysicsextensionsos
+        )
+endif ()
 
 if (USE_PRECOMPILED_HEADERS)
     target_precompile_headers(llgltf REUSE_FROM llprecompiled)
 endif ()
+
+#add unit tests
+if (LL_TESTS)
+    INCLUDE(LLAddBuildTest)
+    SET(llgltf_TEST_SOURCE_FILES
+      llgltfmaterialcore.cpp
+      )
+
+    LL_ADD_PROJECT_UNIT_TESTS(llgltf "${llgltf_TEST_SOURCE_FILES}")
+endif (LL_TESTS)
