+++ indra/llgltf/tests/llgltfmaterialcore_test.cpp
