+++ indra/llprimitive/CMakeLists.txt
@@ -7,18 +7,15 @@ include(LLCommon)
 include(LLCoreHttp)
 include(LLPhysicsExtensions)
 include(LLPrimitive)
-include(GLM)
+include(LLGLTF)
 include(Mikktspace)
 
 set(llprimitive_SOURCE_FILES
     lldaeloader.cpp
-    llgltfmaterial.cpp
     llmaterialid.cpp
     llmaterial.cpp
     llmaterialtable.cpp
     llmediaentry.cpp
-    llmodel.cpp
-    llmodelloader.cpp
     llprimitive.cpp
     llprimtexturelist.cpp
     lltextureanim.cpp
@@ -37,8 +34,6 @@ set(llprimitive_HEADER_FILES
     llmaterialid.h
     llmaterialtable.h
     llmediaentry.h
-    llmodel.h
-    llmodelloader.h
     llprimitive.h
     llprimtexturelist.h
     lllslconstants.h
@@ -66,8 +61,8 @@ target_link_libraries(llprimitive
         llrender
         llappearance
         llmeshoptimizer
+        llgltf
         ll::colladadom
-        ll::glm
         )
 
 if (HAVOK OR HAVOK_TPV)
@@ -90,7 +85,6 @@ if (LL_TESTS)
     SET(llprimitive_TEST_SOURCE_FILES
       llmediaentry.cpp
       llprimitive.cpp
-      llgltfmaterial.cpp
       )
 
     set_property(SOURCE llprimitive.cpp PROPERTY LL_TEST_ADDITIONAL_LIBRARIES llmessage)
